Danish municipalities

A sf dataset providing geometry for the danish municipalities

Usage

municipalities

Format

municipalities

An object of classes sf and data.frame with 99 rows and 9 columns:

code

Unique code

name

The Danish name of the municipality

region_code

The unique code of the region that this municipality is located in

region_name

The name of the region that this municipality is located in

source_changed

When was the source last changed on the data-provider's side

visualcenter_x, visualcenter_y

Visual center of the polygon in WGS84

geometry

Polygons for each municipality

last_update

When was the data in the package last updated by maintainers

Source

Provided by The Danish Agency of Climate Data https://dawadocs.dataforsyningen.dk/dok/om#vilkaar. Terms and conditions apply.